requirements/requirements.txt
fastapi==0.87.0,<0.89.0
fastapi[all]
httpx>=0.23.1,<0.26.0
loguru>=0.6.0,<0.8.0
numpy>=1.22.2,<1.23.0
pandas>=1.4.2,<1.5.0
python-multipart>=0.0.5,<0.8.0
pydantic>=1.10.2,<1.11.0
tox>=3.27.1,<3.29.1

lcd-classification-model==0.0.1
uvicorn>=0.20.0,<0.22.0
